+++ ios/CDVBackgroundGeolocationTests/SQLiteLocationDAOThreadTests.m
@@ -47,7 +47,7 @@
     
     for (int i = 0; i < threadsCount; i++) {
         dispatch_group_async(group, queue, ^{
-            BackgroundLocation *location = [[BackgroundLocation alloc] init];
+            Location *location = [[Location alloc] init];
             location.time = [NSDate dateWithTimeIntervalSince1970:100+i];
             location.accuracy = [NSNumber numberWithDouble:i];
             location.speed = [NSNumber numberWithDouble:32+i];
@@ -71,7 +71,7 @@
     XCTAssertEqual([locations count], threadsCount, @"Number of stored location is %lu expecting %lu", (unsigned long)[locations count], threadsCount);
     
     for (int i = 0; i < threadsCount; i++) {
-        BackgroundLocation *result = [locations objectAtIndex:i];
+        Location *result = [locations objectAtIndex:i];
         XCTAssertTrue([result.time isEqualToDate:[NSDate dateWithTimeIntervalSince1970:100+i]], "time is %@ expecting %@", result.time, [NSDate dateWithTimeIntervalSince1970:100+i]);
         XCTAssertTrue([result.accuracy isEqualToNumber:[NSNumber numberWithDouble:i]], "accuracy is %@ expecting %@", result.accuracy, [NSNumber numberWithDouble:i]);
         XCTAssertTrue([result.speed isEqualToNumber:[NSNumber numberWithDouble:32+i]], "speed is %@ expecting %@", result.speed, [NSNumber numberWithDouble:32+i]);